+++ var/spack/test_repos/spack_repo/builtin_mock/packages/deprecated_tool/package.py
@@ -17,4 +17,4 @@ class DeprecatedTool(Package):
     version("2.0", md5="0123456789abcdef0123456789abcdef")
     version("1.0", md5="fedcba9876543210fedcba9876543210")
 
-    deprecated("@1.0", reason="cve", severity="critical")
+    deprecated("@1.0", reason="vuln", severity="critical")